Improve nullability index for v1 and v2 models
V2 implementations should not look at the box trait ever. There were
also no test cases covering this, so the removal appears to have no
impact on expected behavior.

V1 implementation should look at the default trait and test if it's
set to the zero value for the target type when trying to determine
if a member is boxed. This ensures that models that are round-tripped
and deserialized can still define their same v1 semantics despite the
removal of the box trait (since synthetic box traits aren't
serialized).