Patch synthetic box traits on v2 members
To make v2 models something that v1 implementations can use,
we need to patch box traits onto members if a v1 implementation
should consider the member boxed. This ensures that v1
implementations that are only inspecting the box trait to
determine structure member nullability will continue to function
as intended.

A v2 member is considered synthetically boxed if:
* The member is not explicitly boxed on the member or on the target.
  (this can't happen in an IDL model, but could happen under very
  specific circumstances with a model assembler).
* The member can be boxed because it targets a shape that supports
  the box trait.
* The member has one or both of the clientOptional trait or
  addedDefault trait, or the member member has a default trait that
  is set to the zero value of the targeted type (so that it matches
  v1 default zero value semantics).