Fix round-tripping for prelude shapes
We previously added synthetic box trait to prelude shapes that were
considered nullable in v1 models so that v1 implementations continue
to work as expected. However, when round-tripping models, and loading
a v1 model serialized as a v2 model, we rely on the fact that the box
trait is synthetically added to members rather than target shapes so
that v1 implementations know the member is nullable. However, naive
checks for box traits on the member or the target shape would fail in
this scenario because it would find the box trait on the target shape
though the intent was not to box the member (for example a v1 shape
might be "upgraded" and given a default trait with the zero value,
but when loaded and checked if nullable using box traits, the check
would be incorrect).

To address this, the prelude shapes are no longer given a box trait.
The prelude-1.0 model is also gone and the box trait is moved into
the normal prelude file (note that you still can't use the box trait
in v2 models). The model assembler will now automatically add the
box trait to members that target v1 shapes that had their box traits
removed so that v1 style nullability checks succeed (so long as they
look at members and then target shapes for the box trait).

This change also updates the NullableIndex to understand that when
it is asked if a root level shape has v1 nullability semantics, it
can check if the shape is one of the prelude shapes that had the box
trait and respond with true. However, this special casing is not used
when checking if member targets are boxed because box trait checks
for these kinds of shapes are only safe at the member level.