Keep box traits on root shapes via tag
Ensures that the box trait on root level shapes isn't lost when serializing
a v1 model as a v2 model.

If a root level shape has the box trait, then a "box-v1" tag is added to
the shape so that it is serialized in a 2.0 model. When loading a 2.0 model,
if the "box-v1" tag is found, then a synthetic box trait is added to the
shape.

This is useful for tooling that integrates Smithy models with other
modeling languages that require knowledge of the box trait on root level
shapes.