Enforce private on traits (#1406)
Fixes: #1369

PrivateAccessValidator was not checking trait relationships which allowed private traits to be referenced in any namespace. The HttpConfiguration mixin in aws.protocols was not including private in its localTraits, which caused the private trait to be applied to any shape that mixed in HttpConfiguration. This issue wasn't apparent because trait relationships were not being validated for private access. The private access validation test was updated to check for this.

This could be a breaking change for any Smithy models that are using private traits in other namespaces, either directly or through a mixin that doesn't have private included in its localTraits.

Commits:
* Fix bug where `private` would be applied to every shape that mixes in `HttpConfiguration`

* Fix bug where private access wasn't enforced on trait applications