Warn on unknown trait diffs
In smithy-diff, traits that have `diff` tags applied to their
definitions can be properly evaluated on modification. However,
it is possible for models to be loaded without trait definitions,
meaning no diff evaluation occurs. This commit updates the
ModifiedTrait evaluator to emit events when a trait without a
definition is not constant across the compared models.