Add coverage for `intEnum` in `restJson1` protocol tests
As part of IDL 2.0, the `intEnum` shape was added. The following changes
add several test cases using an `intEnum` shape (and/or aggregates)
under the `restJson1` protocol test suite. Included is a set of shared types,
which target `intEnum` and aggregate-shapes of `intEnum`, and may be used
in follow-up changes to other protocol test suites (awsJson*, *Query, etc.).