Move useIntegerType to jsonschema
The integer type is a sort of vaugely supported thing in json schema,
not jsut OpenAPI. So this moves that configuration up to the generic
settings so it can also be applied to cloudformation and any future
derivative tooling.

The int32 / int64 formats were kept to openapi since that *isn't*
referenced in the the json schema spec.