Use absolute paths to de-dupe model sources
We previously used Paths to de-dupe model sources in the
ModelAssembler, but weren't using absolute paths to
normalize paths before de-duping. This caused issues when
smithy-build.json files defined imports or sources and
sources were provided to the Smithy CLI as positional
arguments. We now normalize files provided to the model
assembler using toAbsolutePath to ensure that relative and
absolute paths are de-duped.

Further, this commit adds the same treatment to SmithyBuild
sources. This isn't needed to de-dupe for the assembler, but
rather to de-dupe files when generating the sources plugin.
Without this, the sources plugin would fail due to conflicting
files that look different because one might be absolute and
another might be relative, but are the same.

Integration test cases were added to the Smithy CLI to test this
because it does not appear reliable to change the current working
directory.