Tweak class caching in node (de)serializers.
Prior to this PR, class caching mechanisms in DefaultNodeDeserializers
and DefaultNodeSerializers would cache classes the first time they were
loaded. If the class was loaded again using a new ClassLoader, and then
retrieved from the cache, there would be a mismatch in the loaded class
and the one retrieved from the cache because class equality is dependent
on the ClassLoader used. This caused class cast exceptions in the language
server.

Based on discussion with Michael offline, this PR introduces a modified
caching mechanism that associates a key-value pair with an instance
of a loaded class. When retrieving a value from the cache, the associated
class is also provided, and the cache will be invalidated if the associated
class is not referentially equal to the currently cached class. This modified
caching mechanism replaces caches in DefaultNodeSerializers and
DefaultNodeDeserializers.

Testing
------
- Test class for the new caching mechanism
- Added test case to ModelAssemblerTest for loading external jars with trait
implementations multiple times. The jar and model used for this test case
was taken from the language server.