Fix grammar for MapMembers
MapMembers grammar was incorrectly stating a BR was required
between the key and value, but actually WS is required. This
was already reflected in the implementation, and the previous
grammar was at odds with being able to use a comma to separate
the key and value members. I added a new test case for 2.0 models
and map parsing.

Closes #1511