Refactor printing to use PrintWriter
Printing from the CLI previously often resulted in hard to read
output because messages were flushed on each new line rather than
as cohesive sections of text. This change updates the CLI to use a
PrintWriter instead, with a buffer, so that large swaths of text are
kept together. Additionally, a new API for creating grouped sections
of text from a CliPrinter were added that include styling methods.