Refactor CLI printing and colors
Printing from the CLI previously often resulted in hard to read
output because messages were flushed on each newline rather than
each log message/write to stdout/stderr, causing contiguous strings of
text that contain newlines (like model validation events) to be
interleaved between threads. This change updates the CLI to use a
PrintWriter instead of a PrintStream to fix this. Additionally, added
a new API for creating grouped sections of text from a CliPrinter that
includes styling methods.

Forcing color/no_color is now done only through environment variables
(FORCE_COLOR, NO_COLOR). This matches lots of other CLI tools, and removes
the clutter of the old options from help output. If the old options are
used, a warning is logged and they are ignored.

The way in which colors are detected and rendered is updated too. This
change introduces an Ansi enum that contains AUTO, FORCE_COLOR, and
NO_COLOR members. AUTO will auto-detect whether colors are supported
using a heuristic, and dispatch to FORCE_COLOR and NO_COLOR based on if
colors are supported. CliPrinter instances have a reference to Ansi and
use it to style text + provide a public API to get the Ansi color setting.

The hueristic used to detect colors is now:
1. Is FORCE_COLOR set? colors
2. Is NO_COLOR set? no colors
3. Is TERM set and set to "dumb"? no colors
4. Is TERM null and the OS is Windows? no colors
5. Is a System console detected (i.e., interactive CLI that isn't
   redirected)? colors
6. no colors

I also updated checkstyle to allow for empty methods with braces on
the same line to accomodate some of the implementations I was added.
I then went and made similar changes to classes in the CLI package
to use this approach.