Add back support for --no-color and --force-color
Per request, adding back support for these arguments. This was
achieved by separating the ColorFormatter from the CliPrinter.
Because CLI arguments aren't fully parsed when bootstrapping the
CLI and there needs to be a resolved ColorFormatter to create an
Environment to give to a comment, I added a ColorFormatter interface
that is implemented by AnsiColorFormatter (formally Ansi). The CLI
will create an instance of a ColorFormatter that defers whether color
is supported or not by creating a custom ColorFormatter that queries
StandardOptions#colorSetting before each write.