Fix dedicated io transform leaving unused shapes (#1419)
This commit changes the way `CreateDedicatedInputOutput` transform checks for
singular references to a shape being used as input or output in an operation.
It now checks to make sure the shape is only referenced by the specified
operation. This fixes an edge case where a shape was being used as both the
input and output for an operation, and would be left unused in the model
after the transformation was applied. Two new test cases were added to
verify the behavior when a shape is used as both input and output.

Fixes: https://github.com/awslabs/smithy/issues/1373