Manually fix some things that are broken
I believe these are instances of things that you parse
but that don't match the grammar. I decided not to diverge
from the grammar in the formatter in those cases because I
think they're not worth it.

On the other hand, I found some other things that you support
that don't align with the grammar but made sense to me so I
tweaked the formatter a bit to follow.