Fix quoted text grammar and parsing (#1535)
Updates Smithy ABNF for QuotedChar and TextBlock, and updates IdlTextParser to
properly parse QuotedChar and TextBlock. Previously, IdlTextParser was allowing
any character to be used as a QuotedChar, despite QuotedChar having certain
restrictions in the grammar. IdlTextParser was updated to follow the grammar's
restrictions properly. The grammar restrictions were also loosened to allow
tab characters in QuotedChar and double quotes in TextBlock, both of which
were already in use. Specifically for double quotes in a TextBlock, one or two
are allowed in a row, and must be followed by a QuotedChar.