Update CLI to ignore dependencies when wrapped by Gradle
When using Gradle, Gradle should handle dependency resolution, not the
CLI. If we detect that Gradle is wrapping the CLI, dependency
resolution is ignored. This helps ensure that previous versions of
the Gradle plugin do not allow dependency resolution to occur, which
would not work well with Gradle and could break users when they upgrade
to a newer version of the Gradle plugin.