Update CLI build and integ tests to work with shading
This commit updates the Smithy CLI build to shade the JAR and still
get the same treatment as other JARs (like adding a license). This
required moving the shadow plugin to the main build.gradle and
selectively disabling it for everything but smithy-cli. Then, when
running the maven publish plugin, we switch between the normal JAR
and the shaded JAR output.

When running integration tests for a pre-release build, the JAR is
not yet in Maven Central, which resulted in integration tests failing.
The same problem occurred when running a classpath based command
(they added built-in dependencies on the current CLI version, but those
versions aren't yet in central). That isn't strictly necessary because
we filter out resolved JARs from provided JARs.

Integration tests for the CLI now use a specific version of Smithy,
1.26.0. The version doesn't really matter since it's ignored when
performing the actual CLI build (since the provided JARs supersede
the given version).