Move CDS warmup to the CLI directly
The Java class data sharing archive can now be created by the CLI
using the warmup command. This makes it possible to build the CDS
archive during an install step so that the archive uses the right
paths. Creating the archive before distributing the CLI results in
the archive using absolute paths pointing to the where the CLI was
built rather than where it is installed on end user machines.