Add client codegen guide to docs
A section under 'Guides' was added to the docs, 'Generating code'.
It covers updating the model from the Quick start guide to
enable codegen, adding the client codegen plugin, and using
the generated client. The example uses the TypeScript code
generator.

A hyperlink target was also added to the implementations doc
for the 'Client code generators' section to make it easier
to refer to in the new guide.