Add labeler workflow for adding github labels (#1593)
This workflow will add the GitHub issue labels in a .github/labels.yml
file. It does not delete our existing labels.

GitHub doesn't seem to have a way to run a workflow on PR merge, so
this workflow just runs on pushes to main. After this PR is merged,
a commit will have to be made into main with the .github/labels.yml
file. After the labeler workflow runs, the workflow file (in this
commit) and the labels.yml file should be deleted.

The labels to be added are the standard labels that AWS SDKs use,
and having these labels will bring consistency across repos.