Clairfy rules for escaping shapes bound to URIs
This updates the rules for escaping values of shapes bound to the
HTTP URI according to the rules of RFC3986. The list of characters to
escape was replaced with a list of safe characters. Initially that
was created by subtracting the reserved characters in section 2.2
from the relevant valid characters defined in the productions for the
particular binding locations, but those all ended up being equivalent
to the unreserved characters production in section 2.3.

closes #1011