Add process based plugins to Smithy build
This commit adds the ability to integrate smithy build with
process based plugins so that the Smithy CLI will be able to
call other CLI tools through a process based plugin model.

To do this, I've introduced the `run` plugin which takes a
`command` array containing the program to run and arguments
to provide. The plugin will open a process and send the
serialized Smithy model to the standard input of the process.
The process can then perform its logic and write files in the
working directory its called with.

A couple things are still missing here that we should figure
out later:

1. The only way to see the output of a plugin is using --debug
   in the CLI. Should there be an output to somehow always
   show it? That will require some kind of hackery in the CLI
   I think.
2. There is no way currently to indicate the process is "serial"
   like how other Smithy Build plugins can. Serial indicates that
   a plugin requires some kind of global locking (like file locks)
   and can't be run in parallel with other plugins.

In addition to adding the run plugin, you can also apply the same
plugin multiple times in a single projectiony by giving it a custom
artifact name using the "artifact-name via plugin-name" syntax.