Remove presence tracking from default values
This commit removes the recommendation for using presence tracking
default values, and instead advocates for serializing default values.
This ensures that whatever an implementation thinks the default value
is currently is unambiguoulsy honored by anything that deserializers
the message later at any point in time, including if the default
value for that member has since changed in the model. This also removes
the large amount of complexity that presence brings with it. While
this does make it harder for a default value to change, changing
defaults is meant to be very rare anyways. So the tradeoff here
seems worth it given changing defaults was and continues to be so
strongly discouraged.