Add --show-traits to select command
The --show-traits option will include the serialize value of
specific traits for each shape matched by the --select command.
The output format is the same as the format used with --vars, but
also included in the output is a 'traits' property that contains a
map of trait shape IDs to trait values. --show-traits is a comma
separated list of trait shape IDs, where omitting a namespace
assumes the trait is in the prelude (e.g., 'required' is the same as
'smithy.api#required'). Matches are not required to have the traits
defined in --show-traits, so if that is the desired filtering, the
selector provided in --selector needs to ensure each shape has the
required traits.

This option can be useful to answer questions like: "what are all the
HTTP headers used in my service?" or "what are the default values of
every shape?"