Improve CLI validation output
This commit adds more colorful and easier to scan validation output
for the CLI.

To support this change, I needed to introduce a new validation event
formatter that uses ANSI codes. Updating the existing formatter would
have added a ton of composition to it, making it hard to understand.

When breaking out a new validation event formatter, I noticed some
duplicate code in handling: 1) creating a human-readable filename
2) loading a line of text from a source location.

To account for (1), I added getHumanReadableFilename() to
ValidationEvent.

To account for (2), I created a new abstraction for grabbing source
context from a SourceLocation.

I ended up needing an abstraction for emitting lots of content that's
wrapped in ANSI color escapes, and I didn't want to add escapes over
and over or have to buffer intermediate output to a string, so I added
a method for wrapping a Consumer<Appendable> that handles adding
ANSI styles before and after the consumer is called.

Style was updated to an interface in case we ever want to support
theming and use HEX code colors.