Use updated output with smithy build and diff
Smithy diff and build now uses the pretty validation output and
color theming options.

Because Styles were copied and repated multiple times, I introduced
a ColorTheme class to store reusable styles. This isn't a swappable
color scheme system, but could be evolved into one in the future if
needed.

For Smithy Diff:

The old and new models are both validated, but only build-failing
errors are shown. When the original model fails, a label is added
before ERROR/DANGER to show it's on the OLD model. When the new
model fails, it shows a NEW label.

When running the diff with two valid models, events have a label
of DIFF. The events shown here use the --severity option and default
to WARNING when not set.

For Smithy Build:

Each projection output is contained in a titled section using a
colored title to match whether the projection passed or failed.
Any events encountered during the projection are emitted using the
updated styling and have a prefix label using the projection name.