Remove hidden --discover option from some commands
The --discover and --discover-classpath options are only necessary
for smithy build and smithy validate to support the Gradle pluing and
the internal build tool within Amazon. This commit removes those
options from ast, diff, migrate, and select, and removes / fixes
tests that relied on them.

We should figure out if we can remove these options from
build/validate or if we're comfortable with them as public options.