Allow IDL serialization order configuration
This commit adds the ability to order the output of the IDL
serializer so that shapes/traits/metadata can be written out in the
order they were originally defined. You can also use alphabetical
ordering or a "preferred" order (the default, and previous behavior).

This change will allow us to potentially use the IDL serializer to
more faithfully serialize transformed Smithy models.