fix error msg for invalid operation io bindings
Fixes issue #1719, which pointed out that error messages when operation
input targeted structures with the `@output` trait (and vice-versa) were
incorrect. The error message was built using the operation <-> target
relationship name for the name of the invalid trait, instead of using
it as the operation property which targeted an invalid shape.

The error message was also updated to include the shape id of the
targeted shape, and existing tests were updated to reflect that
change.