Create Smithy-CLI installation guide (#1697)
Add smithy-cli guide section and install guide

As part of the official Smithy-CLI release, we need a guide on how
to install it for each supported platform. A new guide section is also
added for CLI-specific guides (coming soon). A verification guide
is also written, so users can verify their installation.