Add rudimentary form of error recovery to parser
When we encounter a parser error, we'll now attempt some basic
error recovery by simply skipping token until we reach a token at
the start of a line that is an identifier, documentation comment,
or '@'. This should help with things like LSPs that still should be
able to utilize jump to definition even with a broken or in-progress
model.

Add error recovery to control statements and metadata. Ensure that
invalid identifiers are turned into ERROR tokens instead of throwing.
Error recovery will also consider "$" a recovery token.