Use IntPredicate and inline end of line parsing
This commit makes minor performance improvements to parsing so that
we can use consumeWhile with an IntPredicate rather than needing to
box a char with a Predicate<Character>. To accommodate this, I
deprecated `consumeUntilNoLongerMatches(Predicate<Character>)` in favor
of `consumeWhile(IntPredicate)`. I also added overloads to ParserUtils
to accept int in addition to char, matching how other classes in the
JDK work like Character.isDigit.