Fix inline io parsing with mixins
When there was no space between the `:=` and `with` tokens when
parsing an inline io statement, for example:
```
operation Foo {
  input :=with [Bar] {}
}
```
the parser would skip the `with` and so not realize its parsing a
mixin, failing when it sees a `[` instead of a `{`. This was due
to an extra call to `tokenizer.next()`, which skipped the `with`.
Usually the `next()` would just skip the space before the `with`,
but when the space wasn't present it skipped the `with`.

The inline-io test was updated to include this case.