Add restXml protocol test for http label escaping
Adds an httpRequestTests case to restXml protocol tests to verify
that httpLabel bindings in the request URI are percent-encoded.
restJson1 had an existing test case for this, but was updated to
verify the space character is encoded. An operation was also added
to the testing S3 service model with httpRequestTests to verify the
same thing, but with S3's virtual host addressing. No other protocols
had tests added, because restJson1 and restXml are the only protocols
that support httpLabel bindings.

Changes were tested by generating and running protocol tests in
aws-sdk-go-v2 and aws-sdk-js-v3, confirming the new test case is
generated and runs successfully.