Defer projection failure until after plugins run
This changes `SmithyBuildImpl` to continue applying plugins after one fails,
throwing the error (if present) after they have all completed. This is
useful for example when you want to see the serialized output of a model
you know is valid but a plugin is causing the whole build to fail. A new
test case was added to ensure that artifacts produced by valid plugins
are still created despite the build failing.

This was originally implemented in #1416, but was rolled back in #1429
as a precaution since we had an unrelated issue ocurring at the time.