Fix migrate command version matchers
Previously, Smithy models with versions without a decimal suffix (`.0`)
would not be matched properly in the `smithy migrate` command.

For Smithy models with a version of `"1"`, the upgraded files would
result in conflicting version control statements, e.g.

```smithy
$version: "2.0"

$version: "1"

// ...
```

Similarly for Smithy models with a version of `"2"`, the files would
incorrectly attempt to be upgraded, e.g.

```smithy
$version: "2.0"

$version: "2"

// ...
```

This change loosens the version matchers to match the
[`Version::fromString`](https://github.com/awslabs/smithy/blob/c95943844a65a6623f79ac647e54326d87a68c24/smithy-model/src/main/java/software/amazon/smithy/model/loader/Version.java#L205-L216)
implementation.