Add validation events decorators
Adds a new interface `ValidationEventDecorator` that libraries can
implement and plug into Smithy by making those discoverable using the
java builtin
[`ServiceLoader`](https://docs.oracle.com/javase/8/docs/api/java/util/ServiceLoader.html). All
of the `ValidationEventDecorator` instances are loaded and each
`ValidationEvent` is passed to each of the decorators. The decorators
can make use of the `ValidationEvent#eventId()` method to quickly
filter out the events that it knows how to decorate form the events it
does not care about, and the `ValidationEvent#toBuilder()` to create a
builder that can be used to mutate the instance.

The current use case is to add hints to a know set of events to point
the user into the right solution for it (e.g., you need to pull X or Y
dependency for this unresolved shape). Since often these solutions
might require specific details about the environment in which Smithy
is being used, those decorators cannot be part of Smithy itself but
now can be implemented by libraries and loaded automatically by Smithy
in a similar way in which validators are currently loaded.

For instance, a decorator that suggest the user to use or remove
unused structures might look like this

```java
    @Override
    public ValidationEvent decorate(ValidationEvent ev) {
        if (ev.containsId("UnreferencedShape")) {
            if (ev.getMessage().contains("The structure ")) {
                return ev.toBuilder()
                    .hint("Consider using this structure in your service or remove it from the model")
                    .build();
            }
        }
        return ev;
    }
```