Load the decorators using its own factory
In order to be able to decorate events that are returned before
running the validation logic we need to pull the decoration logic up
to the ModelAssambler itself. For this, a new factory was created to
load the decorators instead of coupling this logic with the
ValidatorFactory one and the assembler will take care of loading them
and passing them down to the validation logic.