Add syntax package to parse and format the IDL
smithy-syntax is used to provide token trees (a very slightly higher
level parse tree) of Smithy IDL models. This token tree is meant to
be lossless, fault-tolerant, and easy to navigate.

The use case for smithy-syntax is to make it easier to perform syntax
analysis in tools like the LSP, IntelliJ, in-place model transforms
(e.g., refactoring tools), and a formatter.